MQTT/requirements.txt
influxdb
paho_mqtt
psutil
python-dotenv
PyYAML
